+++ server/src/main/java/com/example/bank/dto/Payment/PaymentResponseDto.java
@@ -4,14 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 
 import java.time.LocalDateTime;
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class PaymentResponseDto {
     private String paymentId;               // UUID for tracking status
-    private String status;                  // "Completed", "Queued" or "Failed"
+    private String status;                  // "PENDING", "COMPLETED", or "FAILED"
     private Long sourceAccountId;
     private Long targetAccountId;
-    private double amount;
+    private BigDecimal amount;
     private LocalDateTime timestamp;
 }
